+++ packages/evaluate/src/weathergen/evaluate/example_extras/power_spectra/psd_main.py
@@ -22,12 +22,12 @@ for use with the WeatherGenerator model:
 
 OR
 
-./packages/evaluate/src/weathergen/evaluate/example_extras/power_spectra/psd_main.py
+./packages/evaluate/src/weathergen/evaluate/example_extras/power_spectra/psd_main.py \
 --config ./packages/evaluate/src/weathergen/evaluate/example_extras/power_spectra/psd_config.yml
 
 Prerequisties:
 
-Please export the inference into a regular gridded netcdf first using the export package:
+Please export the inference into a regular lat lon gridded netcdf first using the export package:
 e.g.
 uv run export --run-id <INFERENCE_ID> --stream ERA5 \
 --output-dir ../output_nc --format netcdf --regrid-degree 1 \
